Use start and end pointers so all memory can be scrubbed.
In reality, it is almost never or even never required that all memory be
scrubbed because the address space will have memory mapped I/O devices
in it. It is *possible* that x86-based systems use the I/O space for
devices instead of memory mapping any of them, so this code will now
handle that case correctly.