Parameterize MemoryScrubber. (Intermitent failures)
Instead of using traits as generic, use the actual types. This
allows the test code, which has things specific to types,
to get to the specific things it needs to get to.

This fails intermitently even when running one test at a time. How this
happens, I have no idea. It should be completely deterministic. More
work to be done on this.